+++ src/lib/spatial/b-tree.ts
@@ -1,57 +1,5 @@
-import { DataTable, TypedArray } from '../data-table/data-table';
-
-// partition idx indices around the k-th largest element
-const quickselect = (data: TypedArray, idx: Uint32Array, k: number): number => {
-    const valAt = (p: number) => data[idx[p]];
-    const swap = (i: number, j: number) => {
-        const t = idx[i];
-        idx[i] = idx[j];
-        idx[j] = t;
-    };
-
-    const n = idx.length;
-    let l = 0;
-    let r = n - 1;
-
-    while (true) {
-        if (r <= l + 1) {
-            if (r === l + 1 && valAt(r) < valAt(l)) swap(l, r);
-            return idx[k];
-        }
-
-        // Median-of-three pivot selection (using values via idx)
-        const mid = (l + r) >>> 1;
-        swap(mid, l + 1);
-        if (valAt(l) > valAt(r)) swap(l, r);
-        if (valAt(l + 1) > valAt(r)) swap(l + 1, r);
-        if (valAt(l) > valAt(l + 1)) swap(l, l + 1);
-
-        let i = l + 1;
-        let j = r;
-        const pivotIdxVal = valAt(l + 1);
-        const pivotIdx = idx[l + 1];
-
-        // Partition around pivot
-        while (true) {
-            do {
-                i++;
-            } while (i <= r && valAt(i) < pivotIdxVal);
-            do {
-                j--;
-            } while (j >= l && valAt(j) > pivotIdxVal);
-            if (j < i) break;
-            swap(i, j);
-        }
-
-        // Place pivot in its final position
-        idx[l + 1] = idx[j];
-        idx[j] = pivotIdx;
-
-        // Narrow to the side containing k
-        if (j >= k) r = j - 1;
-        if (j <= k) l = i;
-    }
-};
+import { DataTable } from '../data-table/data-table';
+import { quickselect } from '../utils/quickselect';
 
 class Aabb {
     min: number[];
