+++ src/lib/voxel/sparse-voxel-grid.ts
@@ -1,9 +1,6 @@
+import { BlockAccumulator } from './block-accumulator';
 import { MaskStore } from './mask-store';
-import {
-    BlockAccumulator,
-    mortonToXYZ,
-    xyzToMorton
-} from './sparse-octree';
+import { mortonToXYZ, xyzToMorton } from './morton';
 
 const SOLID_LO = 0xFFFFFFFF >>> 0;
 const SOLID_HI = 0xFFFFFFFF >>> 0;
@@ -232,6 +229,75 @@ class SparseVoxelGrid {
             }
         }
         return acc;
+    }
+
+    /**
+     * Get the bounding box of occupied blocks.
+     *
+     * @returns Block coordinate bounds, or null if no blocks are occupied.
+     */
+    getOccupiedBlockBounds(): {
+        minBx: number; minBy: number; minBz: number;
+        maxBx: number; maxBy: number; maxBz: number;
+    } | null {
+        const { nbx, nby } = this;
+        const totalBlocks = this.nbx * this.nby * this.nbz;
+        let minBx = nbx, minBy = nby, minBz = this.nbz;
+        let maxBx = 0, maxBy = 0, maxBz = 0;
+        for (let w = 0; w < this.occupancy.length; w++) {
+            let bits = this.occupancy[w];
+            while (bits) {
+                const bitPos = 31 - Math.clz32(bits & -bits);
+                const blockIdx = w * 32 + bitPos;
+                if (blockIdx >= totalBlocks) break;
+                const bx = blockIdx % nbx;
+                const byBz = (blockIdx / nbx) | 0;
+                const by = byBz % nby;
+                const bz = (byBz / nby) | 0;
+                if (bx < minBx) minBx = bx;
+                if (bx > maxBx) maxBx = bx;
+                if (by < minBy) minBy = by;
+                if (by > maxBy) maxBy = by;
+                if (bz < minBz) minBz = bz;
+                if (bz > maxBz) maxBz = bz;
+                bits &= bits - 1;
+            }
+        }
+        return minBx <= maxBx ? { minBx, minBy, minBz, maxBx, maxBy, maxBz } : null;
+    }
+
+    /**
+     * Find the nearest free (unblocked) voxel to a seed position using
+     * expanding cube shells.
+     *
+     * @param blocked - Grid to search for free cells in.
+     * @param seedIx - Seed voxel X coordinate.
+     * @param seedIy - Seed voxel Y coordinate.
+     * @param seedIz - Seed voxel Z coordinate.
+     * @param maxRadius - Maximum search radius in voxels.
+     * @returns Coordinates of the nearest free voxel, or null.
+     */
+    static findNearestFreeCell(
+        blocked: SparseVoxelGrid,
+        seedIx: number, seedIy: number, seedIz: number,
+        maxRadius: number
+    ): { ix: number; iy: number; iz: number } | null {
+        const { nx, ny, nz } = blocked;
+        for (let r = 1; r <= maxRadius; r++) {
+            for (let dz = -r; dz <= r; dz++) {
+                for (let dy = -r; dy <= r; dy++) {
+                    for (let dx = -r; dx <= r; dx++) {
+                        if (Math.abs(dx) !== r && Math.abs(dy) !== r && Math.abs(dz) !== r) continue;
+                        const ix = seedIx + dx;
+                        const iy = seedIy + dy;
+                        const iz = seedIz + dz;
+                        if (ix < 0 || ix >= nx || iy < 0 || iy >= ny || iz < 0 || iz >= nz) continue;
+                        if (!blocked.getVoxel(ix, iy, iz)) return { ix, iy, iz };
+                    }
+                }
+            }
+        }
+        return null;
     }
 }
